requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
nltk==3.4.3
protobuf==3.8.0
psycopg2-binary==2.8.4
redis==3.3.8
six==1.12.0
timeout-decorator==0.4.1
tqdm==4.35.0
tribool==0.7.3
Werkzeug==0.15.6
